roxydental-ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
statsmodels==0.14.2
python-dotenv==1.0.1
requests==2.32.3